+++ src/tui/tabbed.rs
@@ -3,14 +3,12 @@ use termbox_simple::{Termbox, TB_UNDERLINE};
 
 use config::Colors;
 use config::Style;
+use notifier::Notifier;
 use trie::Trie;
 use tui::messaging::MessagingUI;
 use tui::messaging::Timestamp;
 use tui::MsgTarget;
 use tui::widget::WidgetRet;
-
-use notifier::Notifier;
-use notifier::NotifyFor;
 
 const LEFT_ARROW: char = '<';
 const RIGHT_ARROW: char = '>';
@@ -188,7 +186,7 @@ impl Tabbed {
         }
     }
 
-    fn new_tab(&mut self, idx: usize, src: MsgSource, status: bool, notify_for: NotifyFor) {
+    fn new_tab(&mut self, idx: usize, src: MsgSource, status: bool, notifier: Notifier) {
         use std::collections::HashMap;
 
         let mut switch_keys: HashMap<char, i8> = HashMap::with_capacity(self.tabs.len());
@@ -227,7 +225,7 @@ impl Tabbed {
                 src,
                 style: TabStyle::Normal,
                 switch,
-                notifier: Notifier::init(notify_for)
+                notifier,
             },
         );
     }
@@ -243,7 +241,7 @@ impl Tabbed {
                         serv_name: serv_name.to_owned(),
                     },
                     true,
-                    NotifyFor::Mentions
+                    Notifier::Mentions
                 );
                 Some(tab_idx)
             }
@@ -274,13 +272,13 @@ impl Tabbed {
                     }
                     Some(serv_tab_idx) => {
                         let mut status_val: bool = true;
-                        let mut notify_for: NotifyFor = NotifyFor::Mentions;
+                        let mut notifier = Notifier::Mentions;
                         for tab in self.tabs.iter() {
-                            if let MsgSource::Serv{ serv_name: ref serv_name_ } = tab.src {
+                            if let MsgSource::Serv { serv_name: ref serv_name_ } = tab.src {
                                 if serv_name == serv_name_ {
                                     status_val = tab.widget.get_ignore_state();
-                                    notify_for = tab.notifier.notify_for;
-                                    break
+                                    notifier = tab.notifier;
+                                    break;
                                 }
                             }
                         }
@@ -292,7 +290,7 @@ impl Tabbed {
                                 chan_name: chan_name.to_owned(),
                             },
                             status_val,
-                            notify_for
+                            notifier
                         );
                         if self.active_idx >= tab_idx {
                             self.next_tab();
@@ -334,7 +332,7 @@ impl Tabbed {
                                 nick: nick.to_owned(),
                             },
                             true,
-                            NotifyFor::Messages
+                            Notifier::Messages
                         );
                         if let Some(nick) = self.tabs[tab_idx].widget.get_nick().map(str::to_owned) {
                             self.tabs[tab_idx + 1].widget.set_nick(nick);
@@ -1067,24 +1065,23 @@ impl Tabbed {
         false
     }
 
-    pub fn notify(&mut self, notify_for: NotifyFor, target: &MsgTarget){
+    pub fn set_notifier(&mut self, notifier: Notifier, target: &MsgTarget){
         self.apply_to_target(target, &|tab: &mut Tab, _| {
-            tab.notifier.set_notify_for(notify_for);
+            tab.notifier = notifier;
         });
     }
 
     pub fn show_notify_mode(&mut self, target: &MsgTarget){
         self.apply_to_target(target, &|tab: &mut Tab, _| {
-            let notify_for = tab.notifier.notify_for;
-            if notify_for == NotifyFor::Off {
-                tab.widget.add_client_notify_msg("Notifications are off");
-            }
-            else if notify_for == NotifyFor::Mentions {
-                tab.widget.add_client_notify_msg("Notifications enabled for mentions");
-            }
-            else if notify_for == NotifyFor::Messages {
-                tab.widget.add_client_notify_msg("Notifications enabled for all messages");
-            }
+            let msg = match tab.notifier {
+                Notifier::Off =>
+                    "Notifications are off",
+                Notifier::Mentions =>
+                    "Notifications enabled for mentions",
+                Notifier::Messages =>
+                    "Notifications enabled for all messages",
+            };
+            tab.widget.add_client_notify_msg(msg);
         });
     }